+++ cs/Markdown/Rendering/IMarkdownConverter.cs
@@ -1,0 +1,7 @@
+﻿namespace Rendering
+{
+    public interface IMarkdownConverter
+    {
+        string Convert(string markdown);
+    }
+}
